Improve otslib_get_name error handling
Ensure that the name parameter of otslib_get_name is not NULL, if the
parameter is NULL, there is no point in getting the object name.

It is also possible that the memory allocation to hold the object name
may fail. We must check this and fail the function if it does.

Signed-off-by: Abe Kohandel <abe@electronshepherds.com>